matcher/requirements.txt
fastapi
uvicorn
cloudinary
python-dotenv
pydub>=0.25.1
numpy>=1.22
psycopg2-binary
requests